BUG: persistent_workers must be False if num_workers == 0
When num_workers == 0, the main process is used for data loading.
Using persistent_workers=True in this scenario does not make sense
since there are no workers to persist.

Because the default persistent_workers value in
torch.utils.data.DataLoader is False, it raises an error when
we specify the inconsistent configuration persistent_workers=True
and num_workers=0.

The default timm.data.create_loader behaviour is to use
persistent_workers=True, which creates an error when the default
num_workers value is changed from 1 to 0.

In this commit, we change this behaviour so timm.data.create_loader
will silently change the persistent_workers vlaue to False when
num_workers==0.